Double unique id size, making max IA index 32 bit
It might have been possible for players to send 65535 (2^16-1) input
  actions throughout a session if some system sends like tons and tons
  of tiny input actions. But now with 4294967295 that won't be an issue

+++ Runtime/InputActionSyncAssignerListener.cs
@@ -19,7 +19,7 @@ namespace JanSharp
         {
             InputActionSync inputActionSync = (InputActionSync)(Component)poolObject;
             inputActionSync.lockStep = lockStep;
-            inputActionSync.shiftedPlayerId = ((uint)player.playerId) << LockStep.PlayerIdKeyShift;
+            inputActionSync.shiftedPlayerId = ((ulong)player.playerId) << LockStep.PlayerIdKeyShift;
 
             lockStep.OnInputActionSyncPlayerAssigned(player, inputActionSync);
         }
